R obsolete FileUtilities
fixes #313
